can: remove obsolete get_canfd_dlc() macro
The macro was always used together with can_dlc2len() which sanitizes the
given dlc value on its own.

Link: https://lore.kernel.org/r/20201110101852.1973-4-socketcan@hartkopp.net
(cherry picked from commit cd1124e76d740327be5d8f9ce3785ce1119daf4b)

+++ drivers/net/can/usb/peak_usb/pcan_ucan.h
@@ -226,7 +226,7 @@ static inline int pucan_msg_get_channel(struct pucan_rx_msg *rm)
 }
 
 /* return the dlc value from any received message channel_dlc field */
-static inline int pucan_msg_get_dlc(struct pucan_rx_msg *rm)
+static inline u8 pucan_msg_get_dlc(struct pucan_rx_msg *rm)
 {
 	return rm->channel_dlc >> 4;
 }
